# openapi.yaml
openapi: 3.0.0
info:
  title: whatsapp-multi-gateway API
  description: |
    Production-grade multi-tenant WhatsApp API Gateway + Web Dashboard built on
    Baileys v7. Provides REST endpoints for managing WhatsApp instances, sending
    and receiving messages, HMAC-signed webhooks, real-time Socket.IO events,
    authentication, RBAC and audit logging.
  version: 1.0.0
  license:
    name: MIT
servers:
  - url: http://localhost:3333
    description: Local development server
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/auth/login:
    post:
      operationId: login
      summary: Log in with email and password
      description: Returns a JWT plus the authenticated user and their tenant.
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: owner@example.com
              password: s3cret
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
              example:
                success: true
                data:
                  token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ...
                  user:
                    id: u_9f21ab
                    email: owner@example.com
                    name: Tenant Owner
                    role: owner
                  tenant:
                    id: t_4aa8c0
                    name: Acme Corp
                    slug: acme
                requestId: req_4f9d2c1b
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
              example:
                success: false
                error:
                  code: UNAUTHORIZED
                  message: Invalid email or password
                requestId: req_4f9d2c1b
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/auth/me:
    get:
      operationId: getCurrentUser
      summary: Get the current authenticated user and tenant
      tags: [Auth]
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
              example:
                success: true
                data:
                  user:
                    id: u_9f21ab
                    email: owner@example.com
                    name: Tenant Owner
                    role: owner
                  tenant:
                    id: t_4aa8c0
                    name: Acme Corp
                    slug: acme
                requestId: req_4f9d2c1b
        '401':
          description: Missing or invalid credential
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/auth/api-keys:
    post:
      operationId: createApiKey
      summary: Create an API key
      description: Creates an API key for the caller's tenant. The plaintext key is returned only once.
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateApiKeyRequest'
            example:
              name: production-bot
              scopes: ['*']
      responses:
        '201':
          description: API key created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
              example:
                success: true
                data:
                  id: k_a1b2c3
                  name: production-bot
                  scopes: ['*']
                  keyPrefix: wag_a1b2
                  plaintextKey: wag_a1b2c3d4e5f6...
                  createdAt: '2026-01-01T12:00:00.000Z'
                requestId: req_4f9d2c1b
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Forbidden (role must be owner or admin)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    get:
      operationId: listApiKeys
      summary: List API keys for the tenant
      tags: [Auth]
      responses:
        '200':
          description: List of API keys
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
              example:
                success: true
                data:
                  - id: k_a1b2c3
                    name: production-bot
                    keyPrefix: wag_a1b2
                    scopes: ['*']
                    expiresAt: null
                    lastUsedAt: '2026-01-02T08:00:00.000Z'
                    createdAt: '2026-01-01T12:00:00.000Z'
                requestId: req_4f9d2c1b
  /api/v1/auth/api-keys/{id}:
    delete:
      operationId: deleteApiKey
      summary: Delete an API key
      tags: [Auth]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: API key deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
              example:
                success: true
                data:
                  deleted: true
                requestId: req_4f9d2c1b
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Forbidden (role must be owner or admin)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: API key not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Bearer JWT obtained from POST /api/v1/auth/login. Any protected endpoint
        also accepts an X-API-Key header instead of a Bearer token.
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        API key created via POST /api/v1/auth/api-keys. The plaintext key starts
        with 'wag_' and is shown once at creation. Keys are hashed at rest.
  schemas:
    SuccessEnvelope:
      type: object
      properties:
        success:
          type: boolean
          const: true
        data:
          description: Response payload (shape varies per endpoint)
        requestId:
          type: string
          example: req_4f9d2c1b
      required: [success, data, requestId]
    ErrorEnvelope:
      type: object
      properties:
        success:
          type: boolean
          const: false
        error:
          type: object
          properties:
            code:
              type: string
              example: UNAUTHORIZED
            message:
              type: string
              example: Unauthorized
            details:
              type: object
              additionalProperties: true
          required: [code, message]
        requestId:
          type: string
          example: req_4f9d2c1b
      required: [success, error, requestId]
...
    LoginRequest:
      type: object
      required: [email, password]
      additionalProperties: false
      properties:
        email:
          type: string
          format: email
          example: owner@example.com
        password:
          type: string
          minLength: 1
          example: s3cret
...
    CreateApiKeyRequest:
      type: object
      required: [name]
      additionalProperties: false
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          example: production-bot
        scopes:
          type: array
          items:
            type: string
            minLength: 1
          example: ['*']